Correctly handle unsigned int values returned from TCP_INFO
Motivation:

We used an int[] to store all values that are returned in the struct for TCP_INFO which is not good enough as it uses usigned int values.

Modifications:

- Change int[] to long[] and correctly cast values.

Result:

No more truncated values.

+++ transport-native-epoll/src/main/java/io/netty/channel/epoll/Native.java
@@ -244,7 +244,7 @@ public final class Native {
         tcpInfo0(fd, info.info);
     }
 
-    private static native void tcpInfo0(int fd, int[] array) throws IOException;
+    private static native void tcpInfo0(int fd, long[] array) throws IOException;
 
     public static void setTcpMd5Sig(int fd, InetAddress address, byte[] key) throws IOException {
         final NativeInetAddress a = NativeInetAddress.newInstance(address);
